Make update_depot_tools tell you if it's skipping an update.
I set DEPOT_TOOLS_UPDATE to 0 because I sometimes contribute to
depot_tools (and because I was negatively impacted a few times in the
past by a depot tools update in ways that I can't recall.)

Anyway, if you set the DEPOT_TOOLS_UPDATE variable to 0, then even
explicitly doing `update_depot_tools` from the command line will not
actually do an update, you need to set the environment variable to a
different value before attempting an update.

This script is so old that its behavior probably shouldn't be changed.
But we can tell the user what update_depot_tools is doing and why, and
how to get it to do a real update.

Change-Id: I67458e78b5e64f8ea75ffad73b2ede0955c4d91f